+++ av/stream.pyx
@@ -50,7 +50,19 @@ cdef Stream wrap_stream(Container container, lib.AVStream *c_stream):
 
 cdef class Stream(object):
     """
-    A single stream of audio, video or subtitles within a :class:`Container`.
+    A single stream of audio, video or subtitles within a :class:`.Container`.
+
+    ::
+
+        >>> fh = av.open(video_path)
+        >>> stream = fh.streams.video[0]
+        >>> stream
+        <av.VideoStream #0 h264, yuv420p 1280x720 at 0x...>
+
+    This encapsulates a :class:`.CodecContext`, located at :attr:`Stream.codec_context`.
+    Attribute access is passed through to that context when attributes are missing
+    on the stream itself. E.g. ``stream.options`` will be the options on the
+    context.
     """
 
     def __cinit__(self, name):
@@ -127,6 +139,10 @@ cdef class Stream(object):
         """
         Encode an :class:`.AudioFrame` or :class:`.VideoFrame` and return a list
         of :class:`.Packet`.
+
+        :return: :class:`list` of :class:`.Packet`.
+
+        .. seealso:: This is mostly a passthrough to :meth:`.CodecContext.encode`.
         """
         packets = self.codec_context.encode(frame)
         cdef Packet packet
@@ -139,6 +155,10 @@ cdef class Stream(object):
         """
         Decode a :class:`.Packet` and return a list of :class:`.AudioFrame`
         or :class:`.VideoFrame`.
+
+        :return: :class:`list` of :class:`.Frame` subclasses.
+
+        .. seealso:: This is mostly a passthrough to :meth:`.CodecContext.decode`.
         """
         return self.codec_context.decode(packet)
 
@@ -160,6 +180,7 @@ cdef class Stream(object):
         The format-specific ID of this stream.
 
         :type: int
+
         """
         def __get__(self):
             return self._stream.id
@@ -194,7 +215,8 @@ cdef class Stream(object):
         """
         The unit of time (in fractional seconds) in which timestamps are expressed.
 
-        :type: fractions.Fraction
+        :type: :class:`~fractions.Fraction` or ``None``
+
         """
         def __get__(self):
             return avrational_to_fraction(&self._stream.time_base)
@@ -204,9 +226,14 @@ cdef class Stream(object):
 
     property average_rate:
         """
-        The average frame rate of this stream.
+        The average frame rate of this video stream.
 
-        :type: fractions.Fraction
+        This is calculated when the file is opened by looking at the first
+        few frames and averaging their rate.
+
+        :type: :class:`~fractions.Fraction` or ``None``
+
+
         """
         def __get__(self):
             return avrational_to_fraction(&self._stream.avg_frame_rate)
@@ -215,7 +242,12 @@ cdef class Stream(object):
         """
         The base frame rate of this stream.
 
-        :type: fractions.Fraction
+        This is calculated as the lowest framerate at which the timestamps of
+        frames can be represented accurately. See :ffmpeg:`AVStream.r_frame_rate`
+        for more.
+
+        :type: :class:`~fractions.Fraction` or ``None``
+
         """
         def __get__(self):
             return avrational_to_fraction(&self._stream.r_frame_rate)
@@ -223,7 +255,11 @@ cdef class Stream(object):
     property guessed_rate:
         """The guessed frame rate of this stream.
 
-        :type: fractions.Fraction
+        This is a wrapper around :ffmpeg:`av_guess_frame_rate`, and uses multiple
+        huristics to decide what is "the" frame rate.
+
+        :type: :class:`~fractions.Fraction` or ``None``
+
         """
         def __get__(self):
             # The two NULL arguments aren't used in FFmpeg >= 4.0
@@ -235,9 +271,7 @@ cdef class Stream(object):
         The presentation timestamp in :attr:`time_base` units of the first
         frame in this stream.
 
-        Returns `None` if it is not known.
-
-        :type: int
+        :type: :class:`int` or ``None``
         """
         def __get__(self):
             if self._stream.start_time != lib.AV_NOPTS_VALUE:
@@ -247,9 +281,8 @@ cdef class Stream(object):
         """
         The duration of this stream in :attr:`time_base` units.
 
-        Returns `None` if it is not known.
+        :type: :class:`int` or ``None``
 
-        :type: int
         """
         def __get__(self):
             if self._stream.duration != lib.AV_NOPTS_VALUE:
@@ -259,7 +292,7 @@ cdef class Stream(object):
         """
         The number of frames this stream contains.
 
-        Returns `0` if it is not known.
+        Returns ``0`` if it is not known.
 
         :type: int
         """
@@ -269,9 +302,7 @@ cdef class Stream(object):
         """
         The language of the stream.
 
-        Returns `None` if it is not known.
-
-        :type: str
+        :type: :class:``str`` or ``None``
         """
         def __get__(self):
             return self.metadata.get('language')
@@ -281,7 +312,7 @@ cdef class Stream(object):
         """
         The type of the stream.
 
-        Examples: `'audio'`, `'video'`, `'subtitle'`.
+        Examples: ``'audio'``, ``'video'``, ``'subtitle'``.
 
         :type: str
         """